admin/requirements.txt
Flask==1.1.2
flask_sqlalchemy==2.5.1
flask_migrate==2.7.0
sqlalchemy < 1.4.0
PyMySql==1.0.2
sqlalchemy_utils==0.36.8
flask_jwt_extended==4.1.0
redis==3.5.3
python-dateutil==2.8.1
cryptography==3.4.7